Update Appointment class
